feature(flag database): move flag to its own route
[Finishes # 167280373]
