+++ server/repository/carRepository.js
@@ -90,14 +90,4 @@ export default class CarRepository {
         [new ErrorDetail('Params', 'car id', 'Unable to delete AD', carId)]);
     }
   }
-
-  static saveFlag(userId, carId, flagData) {
-    const flagInfo = [flagData.reason, flagData.description, userId, carId];
-    try {
-      return db.query(createFlag, flagInfo);
-    } catch (error) {
-      throw new ApiError(500, 'Internal Server Error',
-        [new ErrorDetail('Body', 'Flag Data', 'Car could not be flagged', flagInfo)]);
-    }
-  }
 }
